Vehicle: implement gimbal control acquisition
Before commanding a gimbal we need to acquire control of it. If
someone else is in control, we should ask the user if they really want
to take over control or not. For instance the gimbal could be controlled
by a companion computer or the autopilot itself as part of a mission and
a user might not want to interfere.

If no-one else is in control, then we silently acquire control.

This control acquisition only works if the gimbal_manager_status message
is sent out. We might have to request it at a certain rate.

Signed-off-by: Julian Oes <julian@oes.ch>